requirements.txt
Pillow
RPi.GPIO
pyserial
click
cython
werkzeug
pymemcache
